Bugfix if java 8 is used and the KeyStore class is not found (plus a little refactoring to the SSLKeyStore).
